Move all the data listing components into subdirectory
